+++ HISTORY.md
@@ -3,9 +3,13 @@
 
 ## 0.1.0
 
-* `NEW`: first version
+* `NEW`: first version.
 
 ## 0.1.1
 
-* `FEAT`: add `onResize` props
-* `FIXED`: some style issue
+* `FEAT`: add `onResize` props.
+* `FIXED`: some style issue.
+
+## 0.1.2
+
+* `FEAT`: add `collapsed` props.